* master: (75 commits)
  Bumped version to 1.1, updated APPSTORE.md categories.
  Updated app icon. Closes #211.
  Update APPSTORE.md
  Updated release notes for 1.1
  Update APPSTORE.md
  Update APPSTORE.md
  Changed vertical position of social buttons because I just couldn't take it anymore. Closes #236.
  Fix variable name collision warnings.
  Do a cooler map animation on district map load.
  Fix VoiceOver on legislator vote table. Closes #234.
  Added confirmation actionsheet for calling legislator. Closes #232.
  Fix VoiceOver of favorite button
  Remove VoiceOver directions for map that is unusable when VoiceOver is enabled. Closes #235.
  Remove period title from back button. Not sure if this will have unintended consequences. Closes #224.
  Fix VoiceOver reading of hamburger icon. Closes #233.
  Change geolocation marker image. Closes #192.
  Turned on more warnings in build settings and corrected some of the resultant warnings.
  Use congress_app Twitter account instead of SunFoundation
  Updated TestFlight.
  Explicitly create a mutableCopy of NSParagraphStyle in SFTableCell.
  ...

+++ Classes/Controllers/SFDataTableViewController.m
@@ -170,7 +170,7 @@
 
         NSMutableArray *mutableSections = [NSMutableArray arrayWithCapacity:[self.sectionTitles count]];
         [mutableSections addObject:[NSMutableArray array]];
-        for (int i = 1; i < [self.sectionTitles count]; i++) {
+        for (NSUInteger i = 1; i < [self.sectionTitles count]; i++) {
             [mutableSections addObject:[NSMutableArray array]];
         }
